MS data backend for mgf files

The MsBackendMassbank class supports import of MS/MS spectra data from MS/MS spectrum data from Massbank files. After initial import, the full MS data is kept in memory. MsBackendMassbank extends the MsBackendDataFrame() backend directly and supports thus the applyProcessing() function to make data manipulations persistent. The backend does however not support export to mgf files yet.

New objects are created with the MsBackendMassbank function. The backendInitialize method has to be subsequently called to initialize the object and import MS/MS data from (one or more) mgf files. Optional parameter nonStop allows to specify whether the import returns with an error if one of the xml files lacks required data, such as mz and intensity values (default nonStop = FALSE), or whether only affected file(s) is(are) skipped and a warning is shown (nonStop = TRUE). Note that any other error (such as xml import error) will abort import regardless of parameter nonStop.

# S4 method for MsBackendMassbank
backendInitialize(
  object,
  files,
  metaBlocks = metaDataBlocks(),
  nonStop = FALSE,
  ...,
  BPPARAM = bpparam()
)

MsBackendMassbank()

spectraVariableMapping(format = c("Massbank"))

# S4 method for MsBackendMassbank
export(object, x, file = tempfile(), mapping = spectraVariableMapping(), ...)

Arguments

object
Instance of MsBackendMassbank class.

files
character with the (full) file name(s) of the mgf file(s) from which MS/MS data should be imported.

metaBlocks
data.frame data frame indicating which metadata shall be read Default is metaDataBlocks().

nonStop
logical(1) whether import should be stopped if an xml file does not contain all required fields. Defaults to nonStop = FALSE.

...
Currently ignored.

BPPARAM
Parameter object defining the parallel processing setup to import data in parallel. Defaults to BPPARAM = bpparam(). See bpparam() for more information.

format
for spectraVariableMapping: character(1) defining the format to be used. Currently only format = "Massbank" is supported.

x
Spectra() object that should be exported.

file
for export: character(1) defining the output file.

mapping
for export: named character vector allowing to specify how fields from the Massbank file should be renamed. Names are supposed to be the spectra variable name and values of the vector the field names in the Massbank file. See output of spectraVariableMapping() for the expected format.

Author

Michael Witting

Examples

## Create an MsBackendMassbank backend and import data from test xml files.
fls <- dir(system.file("extdata", package = "MsBackendMassbank"),
    full.names = TRUE, pattern = "txt$")
be <- backendInitialize(MsBackendMassbank(), fls)

#> Start data import from 8 files ... 

#> done

be

#> MsBackendMassbank with 10 spectra
#>      msLevel     rtime scanIndex
#>    <integer> <numeric> <integer>
#> 1          2        NA         1
#> 2          2    142.14         1
#> 3          2    142.14         1
#> 4          2    142.14         1
#> 5          2    142.14         1
#> 6          2    142.14         1
#> 7          2    142.14         1
#> 8          2    143.94         1
#> 9          2    143.94         1
#> 10         2    143.94         1
#>  ... 28 more variables/columns.

be$msLevel

#>  [1] 2 2 2 2 2 2 2 2 2 2

be$intensity

#> NumericList of length 10
#> [["intensity"]] 12461 2208 2394 40390 2816 3122 ... 6937 2914 6059 1871 9233
#> [[""]] 646 980 2114 20052 1248 7628 2036 494048 75708
#> [[""]] 10186 142 142 750 138 490 126 ... 14266 1478 1600 16504 1446 109762
#> [[""]] 324 184 138 3770 500 800 7214 3238 ... 2802 206 898 162 166 814 250 1840
#> [["intensity"]] 646 980 2114 20052 1248 7628 2036 494048 75708
#> [["intensity"]] 10186 142 142 750 138 490 ... 14266 1478 1600 16504 1446 109762
#> [["intensity"]] 324 184 138 3770 500 800 7214 ... 206 898 162 166 814 250 1840
#> [["intensity"]] 150 200 32 232 80 12162
#> [["intensity"]] 460 3630 658 74 28 190 44 142 50 52 634
#> [["intensity"]] 282 424

be$mz

#> NumericList of length 10
#> [["mz"]] 84.1 105.1 107.1 114.1 115.1 119.1 ... 393.3 396.3 410.3 411.3 414.3
#> [[""]] 74.0233 132.0807 144.0805 146.0598 ... 170.0597 188.0699 205.0965
#> [[""]] 74.0232 77.0381 86.0027 91.0539 ... 160.0947 170.0596 171.0625 188.07
#> [[""]] 53.0019 53.0383 63.0225 65.0381 ... 158.0817 159.0921 160.0755 170.06
#> [["mz"]] 74.0233 132.0807 144.0805 146.0598 ... 170.0597 188.0699 205.0965
#> [["mz"]] 74.0232 77.0381 86.0027 91.0539 ... 160.0947 170.0596 171.0625 188.07
#> [["mz"]] 53.0019 53.0383 63.0225 65.0381 ... 158.0817 159.0921 160.0755 170.06
#> [["mz"]] 72.0095 116.0517 117.0554 159.0935 186.0558 203.0826
#> [["mz"]] 72.0094 74.0253 116.0511 117.0539 ... 162.0307 186.0548 203.0818
#> [["mz"]] 74.0253 116.0508